+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Strophe.js Change Log
 
 ## Version 1.2.2 - Unreleased
-* Explicitly define AMD modules to prevent errors with AlmondJS and AngularJS.
-* Permit connecting with an alternative authcid.
+* #109 Explicitly define AMD modules to prevent errors with AlmondJS and AngularJS.
+* #111 Fixed IE9 compatibility.
+* #113 Permit connecting with an alternative authcid.
+* #121 Ensure that the node names of HTML elements copied into XHTML are lower case.
 
 ## Version 1.2.1 - 2015-02-22 
 * Rerelease of 1.2.0 but with a semver tag and proper formatting of bower.json