+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 * #109 Explicitly define AMD modules to prevent errors with AlmondJS and AngularJS.
 * #111 Fixed IE9 compatibility.
 * #113 Permit connecting with an alternative authcid.
+* #119 Provide the 'keepalive' option to keep a BOSH session alive across page loads.
 * #121 Ensure that the node names of HTML elements copied into XHTML are lower case.
+* #124 Strophe's Builder will swallow elements if given a blank string as a 'text' parameter.
 
 ## Version 1.2.1 - 2015-02-22 
 * Rerelease of 1.2.0 but with a semver tag and proper formatting of bower.json